Add files via upload
Axial Gun Control: This script allows you to have a user controlled extra mg 
CameraPositionManager: This script allows you to switch between different cameras and activate / deactivate different objects
GunnerControl: This allows a gunner camera to use a thermal vision script ( i didnt put it tho ) and zoom in 
PlayerBarrelgun: This allows you to shoot the main gun according to a given rigidbody ( shell ) and a force (for setting the shells velocity)
PlayerMG: this allows you to have a machine gun. uses similar code to the playerbarrelgun script
PlayerTankDrive: This allows you to drive a tank. A speed control is added allowing you to speed up and speed down instead of teleporting
PlayerTurretScripts: This script allows you to rotate your turret and your gun barrel. It can set a limit to how much you can elevate etc.
ShellScript : this script enables your shells to destroy after time and penetrate armor
smoke grenade: this allows you to make a smoke grenade which plays a particle system on collision
smoke grenade script: this allows you to dispense a smoke grenaed
soundmng: this is a script that controls the sounds played by the guns etc.
tankhealth: this is a script that enables armor for enemies ( the object needs a "Enemy" tag btw and the shell a "Boolet" tag)
Tankselectmanager allows you to select a tank via the UI. if the selected tank is a bruh industries tank it will enable the furry enemy and vice versa. This can also disable the UI to clear up the screen. there are no comments on scripts so u have to figure it out.